engine/llm_manager.py
import os
import openai
import instructor
from dataclasses import dataclass
from typing import Optional

# ---------------------------------------------------------
# Scaffolding: Classes and functions mentioned in the prompt
# ---------------------------------------------------------
class TokenOptimizer:
    pass

def get_cache():
    return {}

@dataclass
class ModelConfig:
    model: str
    max_tokens: int
    cache_ttl: int

# ---------------------------------------------------------
# Routing Table: Configured for DeepSeek models
# ---------------------------------------------------------
ROUTING_TABLE: dict[str, ModelConfig] = {
    "triage": ModelConfig(model="deepseek-chat", max_tokens=256, cache_ttl=300),
    "simple_qa": ModelConfig(model="deepseek-chat", max_tokens=512, cache_ttl=86_400),
    "code_generation": ModelConfig(model="deepseek-coder", max_tokens=4_096, cache_ttl=3_600),
    "compliance_map": ModelConfig(model="deepseek-chat", max_tokens=2_048, cache_ttl=604_800),
    "report_gen": ModelConfig(model="deepseek-chat", max_tokens=8_192, cache_ttl=86_400),
}

# ---------------------------------------------------------
# Main LLM Manager
# ---------------------------------------------------------
class LLMManager:
    def __init__(self, optimizer: Optional[TokenOptimizer] = None, spec_path: str = "PROJECT_SPEC.md") -> None:
        api_key = os.environ.get("DEEP_SEEK_API_KEY")
        self._raw_client = openai.OpenAI(
            api_key=api_key,
            base_url="https://api.deepseek.com" # هذا هو رابط ديب سيك
        )
        self._instructor_client = instructor.from_openai(self._raw_client) # تحويل لـ OpenAI
        self._cache = get_cache()
        self._optimizer = optimizer or TokenOptimizer()
        self._spec_path = spec_path

    # يمكن إضافة دوال إضافية هنا لاحقاً للتفاعل مع النماذج
